requirements.txt
flask
flask-cors
python-dotenv
requests
PyPDF2
gunicorn
groq
flask-limiter
werkzeug
packaging==24.2
pytesseract
Pillow
sentence-transformers
faiss-cpu
numpy